+++ tests/core/test_save_restore.py
@@ -22,10 +22,10 @@ import torch
 from omegaconf import DictConfig, OmegaConf, open_dict
 
 from nemo.collections.asr.models import EncDecCTCModel, EncDecCTCModelBPE
-from nemo.collections.common.models import MockModel
 from nemo.collections.nlp.models import PunctuationCapitalizationModel, TransformerLMModel
 from nemo.core.classes import ModelPT
 from nemo.utils.app_state import AppState
+from nemo.utils.models import MockModel
 
 
 def getattr2(object, attr):
